# Conflicts:
#	packages/core_module/lib/src/events/components/stores/create_event_store.dart

+++ packages/service_module/lib/src/ui/views/services_collection_view.dart
@@ -1,7 +1,6 @@
-import 'dart:io';
+//import 'dart:io';
 
 import 'package:core_module/core_module.dart';
-import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 
@@ -143,15 +142,18 @@ class _ServicesCollectionViewState extends State<ServicesCollectionView>
                                                 icon: AppIcons.trash,
                                                 label: 'Deletar',
                                                 action: () async {
-                                                  _bloc.add(
-                                                    DeleteItemEvent(
-                                                      index: index,
-                                                      context: context,
-                                                    ),
-                                                  );
-                                                  if (context.mounted) {
-                                                    pop(context);
-                                                  }
+                                                    await showConfirmationDialog(
+                                                      confirmAction: () async {
+                                                        _bloc.add(
+                                                          DeleteItemEvent(
+                                                            index: index,
+                                                            context: context,
+                                                          ),
+                                                        );
+                                                      },
+                                                      title: "Deletar Culto",
+                                                      message: "O culto será deletado permanentemente. Tem certeza?",
+                                                      context: context,);
                                                 },
                                               ),
                                             ],
@@ -225,7 +227,7 @@ class _ServicesCollectionViewState extends State<ServicesCollectionView>
                                                 right: 16,
                                               ),
                                               child: IconButtonWidget(
-                                                sizeIcon: ResponsivityUtil<double>(sm: 23, xl: 26).get(context),
+                                                sizeIcon: ResponsivityUtil<double>(sm: 21, xl: 23).get(context),
                                                 color: AppColors.darkGreen,
                                                 iconFormat: IconFormat.svg,
                                                 iconPath: AppIcons.arrowForward,
